Cleanup: Remove uneccessary Javadoc comments, rename Data.
